+++ src/environment.ts
@@ -1,5 +1,6 @@
 
 type Env = {
+  pkgPortable: string,
   production: boolean,
   buildTime: string,
   gitVersion: string,
@@ -7,6 +8,7 @@ type Env = {
 
 declare var WOURNAL_ENV: Env;
 const environment = {
+  pkgPortable: WOURNAL_ENV.pkgPortable,
   production: WOURNAL_ENV.production,
   buildTime: WOURNAL_ENV.buildTime,
   gitVersion: WOURNAL_ENV.gitVersion,