+++ src/renderer/app/debug-info.ts
@@ -2,7 +2,7 @@ import { Component, h, rx } from "@mvui/core";
 import { OpenDialog } from "common/dialog-manager";
 import * as ui5 from '@mvui/ui5';
 import { getLogger, getLogHistoryText } from "util/Logging";
-import environment from "environment";
+import environment from "Shared/environment";
 import PackageJson from 'PackageJson';
 
 const LOG = getLogger(__filename);