+++ webpack.main.config.ts
@@ -1,7 +1,7 @@
 import type { Configuration } from 'webpack';
-
 import { rules } from './webpack.rules';
 import { plugins } from './webpack.plugins';
+import path from 'path';
 
 export const mainConfig: Configuration = {
   /**
@@ -16,5 +16,9 @@ export const mainConfig: Configuration = {
   plugins,
   resolve: {
     extensions: ['.js', '.ts', '.jsx', '.tsx', '.css', '.json'],
+    alias: {
+      PackageJson: path.resolve(__dirname, './package.json'),
+      Shared: path.resolve(__dirname, './src'),
+    },
   },
 };