+++ webpack.plugins.ts
@@ -10,6 +10,7 @@ export const ForkTsCheckerWebpackPlugin: typeof IForkTsCheckerWebpackPlugin =
 
 export const plugins = [
   new DefinePlugin({
+    'WOURNAL_ENV.pkgPortable': process.env.PKG_PORTABLE === 'true',
     'WOURNAL_ENV.production': process.env.NODE_ENV !== 'development',
     'WOURNAL_ENV.gitVersion': JSON.stringify(git('describe --always')),
     'WOURNAL_ENV.buildTime': JSON.stringify(new Date().toISOString()),